Monumental Changes
Alpha.html was given a contact form that will soon be moved to contact.html. The logo in the header was fixed and other minor changes were made. After that, Alpha.html was RENAMED to index.html and the current index.html was renamed delta.html. CSS was given to match these visual changes but, the names of the CSS files stayed the same. The README was updated to include the new tools used in this 3rd redesign of the website and a screenshot was added.
That's all!